Extract Slater--Condon rules for specific matrix elements.
In FCIQMC we know the excitation level, so want to be able to
call the procedures for specific excitation levels directly.

Only those that are immediately useful are extracted: evaluating the
matrix element <D|H|D_ij^ab> for the Hubbard model in momentum space
is even easier as in FCIQMC we have created excitations only when
the connecting integrals are allowed by symmetry.