Use modulo rather than mod to determine walker processor id.
This is because modulo(x,nprocs) is always positive, whereas
mod(x,nprocs) isn't (as x can be positive or negative).  This is a
result of the hash being an unsigned integer in C, which is converted
into a signed integer in fortran.

Also, allow fciqmc to be performed on multiple processors (currently
broken).