Refactor.
If a spawning is successful it now returns the number of particles to
spawn and the connection information.  This will allow the initiator to
use the same spawning algorithms, as only setting the information about
the spawned walker will change (as information on the parent determinant
will need to be included).