Change from doing scatterv in place to using a different receive buffer.
Doing the scattering in place seemed to cause memory issues with ifort
on cmth and cuc3 workstations, though a simple example worked fine.
Array bounds checking didn't reveal a problem in the hubbard code though
and the arguments to scatterv seemed correct.

Doing the scattering into a different array (and we already have
declared the memory to use elsewhere!) is simpler and works round this
issue.