Change find_permuation_excitx (x=1,2) to use bit strings.
Previously the list of occupied orbitals was used to calculate the
permuation required to align two determinants given a determinant and
the excitation connecting it to another determinant.

This was inconvenient.  An efficient ct recursive algorithm cannot
provide a sorted list of occupied orbitals efficiently, and this is
required for these procedures.  Happily, a (almost certainly faster)
bitwise algorithm is possible.  find_permutation_excit2 is extensively
commented (find_permutation_excit1 is trivial given
find_permutation_excit2).