Workaround for weird problems with intel compiler.
It appears that the intel compilers might have a problem with the (:end)
syntax for denoting array slices.  I found seg faults on the U=2 14s
real space calculations in the quicksort routine after 4220 iterations.
It might represent a memory problem though.  Need to investigate further...

gfortran doesn't have a problem with the code...