Distribute occ_list0 correctly.
The Heisenberg model does not know nel until it is worked out from ms_in
and the number of sites, and so nel cannot be used before that point, as it
will either be 0 (if the compiler is kind) or contain garbage.

Also check that the supplied reference det is the correct size.