Bug fixes to freezing in molecular systems.
I have assumed that

i)  the basis functions in FCIDUMP files are ordered by their
    single-particle eigenvalues
ii) the single-particle eigenvalues are present.

Sadly this is not always true--e.g. MOLPRO output---so still some work
to do.  *sigh* Nevertheless, it works with QChem FCIDUMP files and we
now reproduce FCI energies correctly.