Correct conversion to spin orbitals in molecular excitation generators.
Molecular systems have the down spins as even basis functions; this was
not correctly taken into account when converting a spatial label into
the appropriate spin function.

UHF and RHF QMC calculations now appear to be working.

I really need to do a careful audit of the code and ensure that the
comments/variable names are consistent with the spin they claim to refer
to, at least for sanity's sake!