Split spawning routines into system-specific modules.
This is to help with maintenance and being able to find the relevant
code quickly.  It is also the first step on the reorganisation of the
spawning event, in which the top-level spawning routines are general
rather than system-specific.  Currently too many of them look like:

    call gen_excit_*(...)
    call attempt_to_spawn(...)

with simply different comments.  This makes implementing, say,
importance sampling particlarly painful, as so many routines would have
to be changed rather than simply creating a new spawning routine which does

    call gen_excit_*(...)
    call importance_sampling(...) ! adjusts hmatel accordingly
    call attempt_to_spawn(...)

Furthermore, the current structure predates the folded spectrum,
continuous time and CCMC algorithms, all of which need access to the
excitation generators rather than the top-level spawning routines.  This
has resulted in a general trend for the spawning routines to become more
similar as the actual excitation generation logic has been re-factored
out of them.  We should use this logical progression in a more logical
fashion. :-)