population argument to update_proj_energy_* is now real.
In CCMC the amplitudes must be normalised and hence are real (rather
than integers in FCIQMC and DMQMC).  Handle this by requiring the
population to be passed in as real.  This avoids passing in an
additional parameter (the normalisation factor) and doing
a division/multiplication at the cost of a type cast.  Should be
a minimal impact on FCIQMC and DMQMC.