Change to initiator algorithm.
We now handle multiple spawning events independent of the order they come in.
This is subtley different for 3 or more spawning events from the original
algorithm in the original i-FCIQMC paper, but is more satisfactory and
apparently similar to what is now used in the Alavi code.  The case where
non-initiator determinants spawn more psips of the opposite sign to those
produced from initiator determinants is also handled better (we now only keep
them if the non-initiator psips came from multiple coherent events).

Exact results change but these are really quite minor algorithmic changes so
the results should not within stochastic error.