CCMC bug fix: correct sign for the death step.
I realised that, in a similar fashion to the spawning terms, there must
be a (potential) sign change in the matrix elements for the death/clone
step.  This is not currently evaluated in an efficient manner...

CCMC is now working, at least for full CC in a Hilbert space
of 6 determinants.  The projected estimator is currently incorrect as
the probability of generating a cluster is not (yet) taken into account.

I hate signs.