Treat HF particles cloned from H particles on the same level as non-diagonal spawning events.
This takes into account the case where a H particle spawns an HF
particle on the same determinant but that determinant does not have
a non-zero HF population.  Hence such events can not be handled
'in-place' with the initiator approximation.

From the comments it looks like I thought of this when originally
writing the code, but did not actually implement this!

HFS with the initiator approximation is currently giving biased
expectation values however.