Avoid using import in interfaces defined in subroutines.
This is strictly a sop to compatiblity with gcc 4.6, which appears to
not accept the import statement in interfaces that aren't defined at the
module level (as done in several places elsewhere in the codebase).
This bug is fixed in gcc 4.7.2 (if not earlier), but we'll work round it
so we can compile on slightly older machines.

Evidently there are disadvantages to developing under Arch Linux with
all the latest goodies installed!