Routine to calculate the cumulative population on determinants/excitors.
We define the cumulative population to be the total population on
a given determinant/excitor and all determinants/excitors preceding it.
This is completely arbitary as the ordering of excitors and determinants
is arbitrary.  It is, however, useful if we wish to seek out a specific
excitor/determinant by its population relative to the total population.