Select excitors with weight proportional to population.
Generate a cluster from the set of excitors where each excitor in the
cluster is selected with probability proportional to its population.

This matches Alex Thom's original implementation and, more importantly,
is far more stable: a uniform selection (as previously implemented)
gives rise to rather small selection probabilities and hence population
explosions if an excitor with a large population is chosen.