Use reals for the cluster population when creating a cluster of excitors.
Previously we were using integers.  However, as we don't normalise the
population until the end of the cluster creation, intermediate values of
the population can overflow a standard integer.  This changes exact
results (as numerical noise results in changes to the precise Markov
chain followed), but results should not be affected within error bars.