BUG FIX: detect local changes in Makefile correctly.
Previously we were doing things like:

VCS_LOCAL_CHANGES := (shell git diff -- $(SRCDIRS) ... )

After switching to the generic_makefile, this line is in make.inc which
is included in the Makefile *before* SRCDIRS is defined.  As VCS_CHANGES
was being simply expanded, SRCDIRS was blank and so VCS_LOCAL_CHANGES
was defined if *any* changes were made to the repository.  The solution
is to allow VCS_LOCAL_CHANGES to be recursively expanded so SRCDIRS is
correctly set in the git command.