Make it easier to run blocking.py without having an X connection.
When pylab is imported it attempts to connecting to the X server.  If this
can't happen (e.g. over a standard ssh connection) then it throws
a RuntimeError (not an ImportError).  Handling this exception makes it easier
to use blocking.py remotely, especially given Will's work on automatically
detecting the most appropriate block.