Refactor updating projected estimator procedures.
Projected estimator routines now take all the necessary information as
arguments and return some useful information (i.e. the excitation
connecting the determinant to the reference, matrix element) which can
be reused for evaluating Hellman-Feynmann quantities.

This is a step towards having code that is free of any global state (a
long-term goal).

More importantly for the immediate future, it is a step towards adding
operators for Hellman-Feynmann sampling without having to copy the
updating of the standard Hamiltonian (energy) projected estimator for
each different operator.