CCMC optimisation: faster cluster selection.
Rather than selecting one excitor at a time, choose all excitors in
a cluster by their population and then form the cluster.  This allows us
to do a sort on the population array (small and so fast) and hence
enables us to choose progressive tighter limits on the binary search of
the cumulative population array, which is much faster.  A quick test
(10000 iterations of CCSDTQ Ne, cc-pVDZ basis) shows this small change
gives a speedup of ~7%.  Not bad...

Note: CCMC results change (but not beyond error bars) as we now draw
slightly more random numbers.  Previously we only select an excitor in
a cluster until the cluster size was met or the cluster was found to be
forbidden.  Now we generate all excitors in the cluster and then check
if it's allowed.