Use a hash table for creating RDM from the full density matrix.
Many elements of the full density matrix contribute to the same reduced
density matrix element.  If we create the list of RDM elements directly
from the full density matrix, we can re-use the annihilation
infrastructure to sort and then add the number of particles on the same
RDM element.  The downsides are that the same storage as the full
density matrix is required and (worse) quicksort performs very, very
badly if there are a large number of repeated elements in the array to
be sorted (especially if the number of unique elements is comparatively
small).

Alternative approach: use a hash table to create the contribution to the
RDM from particles on each processor, such that (pre-communication) each
RDM element occurs at most once per processor.  This is *substantially*
faster (by a factor of 10 overall in one test), especially if the RDM is
small compared to the full DM.

WARNING: this requires the hash table to be reset every iteration.  As
such, it is vital that spawned_rdm_length is as small as possible in
order to get good performance.  We could address this by adding code to
dynamically resize the hash table as required...