Variable algorithm for assigning bit strings to processors.
In CCMC we need to be able to move excitors around.  So, instead of
using the algorithm p = modulo(hash(s),n_p), where s is the bit string,
n_p is the number of processors and p is the processor which owns s, we
incorporate a value which depends upon the simulation time (e.g. the
iteration number).  WARNING: if used, then existing particles must be
redistributed...