Updated the config files for cx2.
Can now use the 2013sp1 intel-suite + a version of hdf5 has been
built with fortran support into modules enviroment.
module load hdf5/1.8.12 needs to be added to jobs script.
Paths to the hdf5 dir had to be hard coded as modules is not set
up to export an enviromental variable HF5_HOME as it is for cx1
it simply adds /apps/hdf5/1.8.12/bin to PATH.