Bug fix: generate determinant of desired level in truncated Hilbert space MC.
The algorithm for sampling the size of a truncated Hilbert space
required generating a random determinant at a fixed excitation level.
Unfortunately, just doing N moves is not enough to guarantee an
excitation level of N from the reference as we could excite into a newly
created hole (e.g. the excitation 2->12, 6->16 and 4->6 would be
a double excitation rather than triple excitation).  This meant we
generated too many determinants (more than we thought!) at lower
truncation levels and hence got biased results..

The (very simple) solution is to ensure we only excite into orbitals
which are not occupied in the reference.  We were almost there in making
sure we didn't have excitations like 2->2...

Reported by: Will Vigor.