Fix output bug in load balancing report.
We told MPI we were sending one data type and actually passing (and
receiving) another.  All hell broke loose as a result.  Oops.  As a result, garbage was received and the load balancing report frequently looked like:

Min # of determinants on a processor:     869814
Max # of determinants on a processor:   ********
Mean # of determinants on a processor:  ***********

instead of

Min # of determinants on a processor:     869814
Max # of determinants on a processor:     869814
Mean # of determinants on a processor:    869814.00

assuming perfect load balancing (Fionn can but dream!).