Correct Hubbard Coulomb integral if spin is not conserved.
.eqv. has a lower order of precedence than .and. and hence .not.a .and.
a.eqv.b is interpreted as (.not.a .and. a) .eqv. b which is true if b is
false, where a (b) is true if the ij (kl) pair in the integral <ij|kl>
are of the same spin.  As the Hubbard Coulomb integral in momentum space
is only non-zero for integrals involving anti-parallel spins, this only
caused the wrong result if an integral was requested in which spin was
not conserved.  Fortunately, integrals are currently only requested if
spin is conserved.  Still, this behaviour is unexpected and doesn't
match the claims made in the interface comments.