Bugfix: hybrid parallelisation and spawning.
We were not using the first block of a processor block in the spawn_t object
correctly.  We were using

head_start(0,iproc)+nthreads

(which is correct) but using it to find the number of elements stored in a block using

head(0,iproc)-head_start(0,iproc)+nthreads-1

rather than

head(0,iproc)-(head_start(0,iproc)+nthreads-1)

or the simpler

head(0,iproc)-head_start(nthreads-1,iproc).  This caused the garbage to be sent to
processors, which led to propagation of incorrect information and hence chaos.