Correct spawning rate.
Once upon a time, the spawning rate was (I think!) calculated before annihilation.  (This
was done in certain older algorithms.)  The combination of repeated code into end_mc_cycle
meant the spawning rate was moved to being calculated after annihilation.  This caused the
number of successful spawning events to be the number of successful spawning events from
all processors onto unoccupied determinants stored on the processor rather than the number
of successful spawning attempts made by psips on the processor.  The easiest fix is to
(optionally) calculate the number of successful events just prior to any communication of
the spawned particles in the annihilation process.