Make ScaLAPACK happy if a processor does not have part of a matrix.
If the distributed matrix is small, then some processors might contain
no elements of the matrix.  ScaLAPACK can handle this if:

* descinit (as documented in its interface) gets max(1,nrows) or
  max(1,ncols), as appropriate, where nrows (ncols) is the number of
  rows (columns) owned by the processor;
* routines are called on all processors, even those with no elements.

Before we were passing nrows and ncols to descinit (even if 0) and not
calling ScaLAPACK routines from processors which did not possess any
elements.  Both of these caused ScaLAPACK to hang indefinitely...oops!