Don't test spawning rates (R_spawn).  TEMPORARY.
This is a temporary change.  A bug fix (15e1a07b8) to how the spawning
rate was calculated changed this value slightly, which we don't actually
care about too much.  It is a useful check but it wasn't checked with
the old extraction script and some of the current benchmarks date back
to before this fix.  Once the benchmarks have been regenerated, we
should check R_spawn again.