Reuse the trial density matrix.
Instead of generating a new trial density matrix from
scratch each beta loop add option to reuse the guess from
the previous beta loop and perform the metropolis algorithm
on top of this. This is much faster for large
populations / metropolis attempts but also correlates results
across beta loops. Remains to be seen if the gain isn't overshadowed
by the (potential) increase in the size of the error bars when
blocking analysis is performed.