Be more forgiving when plotting blocking analysis.
If no information for the error in the standard error is supplied, then
regard it as 0.  This is convenient if we're plotting data from
pyblock.error procedures which combine data items. Such procedures just
return the standard error and not the error associated with it (which
isn't particularly meaningful/helpful in the blocking analysis).