Correct cluster selection probability for linked CC and MPI.
Previously we weighted the probability of selecting a cluster of X
excitors by 1/nprocs^{X-1}.

By definition an excitor is always on the same processor as itself.

In unlinked CC this is ok, as a cluster containing an excitor more than
once is discarded.  However, in linked CC a cluster can contain the same
excitor multiple times.  We hence need to weight the probability of
selecting a cluster of X excitors by 1/nprocs{X'-1}, where X' is the
number of *unique* excitors selected.