Treat CC death for reals on the same footing as spawning.
In FCIQMC death occurs for particles which already exist and so it makes
sense to do it exactly.  In CCMC we select clusters of excitors, the
combination of which may not yet exist and so death (at least for
clusters of more than one excitor) is really a spawning event.
This is compounded by (currently) calling death once per excip rather
than once per excitor (which would be possible for the non-composite
approach).

For clusters which we select a lot (i.e. the reference and individual
excitors which are highly occupied), this leads to a large number of
entries in the spawned array each with a small weight, which is not
ideal for communication/annihilation.  Stochastically rounding this to
the spawn cutoff, as is done for spawning in FCIQMC and CCMC,
essentially solves this problem.