Be more flexible in the length of the SHA1 hash passed to init_calc_defaults.
When compiled inside a git repository, make passes in the git sha1 hash,
which is of length 40.  If outside a git repository, a warning message
shorter than 40 is passed in, which was causing a character length
mismatch at compile time.  Instead allow any length to be passed in but
document that only the first 40 characters in the string are used...