Change half density matrix so that it doesn't use weighted sampling.
The half density matrix code used to use the weighted importance sampling to
reduce spawning away from the diagonal by a factor of 2.0, then, when spawns
below the diagonal are reflected to above the diagonal, the off-diagonal
elements won't be twice as large, because half as many spawns have been created.
Then, spawns to the diagonal are increased by a factor of 2.0 to account for the
fact that only half as many spawning attempts are made.

I find this dependence on the importance sampling rather undesirable (especially
since there is new importance sampling work being done). Therefore, the half
density matrix approach no longer uses the weighted sampling. Instead, the
elements are just allowed to be twice as large as they would otherwise be.
Although there are no psips below the diagonal, the extra psips above the
diagonal exactly account for the psips that should be below the diagonal, so
it all works out OK.