Correct MPI timing output
Correct MPI timing output, which was previously just printing 0.0.
Also, it is important to look at both the synchronisation time as
well as the actual communication time. I have therefore made a
derived type for holding both, and an object of this type is held by
semi_stoch_t and spawn_t types. Therefore, both semi-stochastic and
spawned MPI timings are now considered.