Remove folded spectrum extension to FCIQMC.
Folded spectrum FCIQMC turned out not to be great--the reduction in
sparsity of the Hamiltonian and compression of the spectrum required the
use of very small time steps and made it hard to converge to the desired
state.  It turns out that it's hard/impossible to go beyond what
conventional FCI is capable of with this approach and there are more
promising avenues for excited states.