Update load balancing benchmark and add new test.
Reset the benchmark for the load balancing test, which needed doing
after the recent changes to perform redistribution of determinants
at the start of an iteration (instead of at the end). Also, create
a benchmark for semi-stochastic+load balancing, which also uses the
new option to turn semi-stochastic on relative to the shift starting
to vary.