Import AOTUS library (MIT license).
I have only imported the files we'll actually need.  See aotus/README.md
for details and upstream info.  Further, we are using the last version
for lua 5.2.  lua 5.3 includes a few incompatible changes and is not
widely available yet (ie not even the default on Arch Linux!).

Note: this requires linking to the lua library.  Config files have not
yet been updated...