BUG FIX (long-standing): correct initialisation in DMQMC.
tot_walkers (the head of the main particles lists) was not set in DMQMC,
leading to memory accesses if it was not zeroed for us.  Not sure how
this hadn't caused problems before!

Results are unchanged as it is initialised later correctly (but not
before it is used, leading to a seg fault in some cases).