Remove CAS option except for read_in system.
The CAS functionality is only implemented for read_in system.
Implementing it for the UEG would be straightforward(ish).  Doing so the
hubbard_k harder (due to the assumption that a close symmetry group is
in use) and somewhat meaningless for the other systems, which don't have
a good reference.

ps That we can't do

arr(:) = [ ... ]

is rage inducing.  (But not as rage inducing as incorrect dimensions not
being picked up by all compilers.)