Change murmurhash_bit_string to work in bits rather than array elements.
Instead of hashing the entire array passed in, murmurhash_bit_string now
hashes the smallest multiple of 32 bits which contains the number of
bits it is told to hash.  This will allow for bit strings to later be
hashed to the same value, irrespective of whether the 64-bit version
contains additional padding at the end of the string.