Move doing_exact_rdm_eigv out of dmqmc_rdm_in_t type.
Instead, move it to calc for now, grouped with other FCI options.
I don't want to have to pass a DMQMC input type into the FCI routine.
It seems more natural to have this logical with other FCI data, which
should probably be turned into a type later.

The setup_rdm_arrays routine, used by FCI RDM calculations, will also
have to be updated to allow it to be called from FCI calculations
without passing in a DMQMC RDM type.