Simplify application of the offset in assigning states to a processor.
Given a significantly good hash function, we really only need to modify
one bit in the bit string.  Given we exploit this for distribution
across processors anyway, it is safe to reuse this for changing the
processor location.  We do this by applying the offset (via an ieor) to
only the lowest (at most) 31 bits, which neatly makes it identical for
different DET_SIZEs.