Work around subtle code gen issue with Intel compilers.
Following the FCI purity (5df401e), intel compilers (at least the 2013
and 2015 versions I tested) incorrectly generated code for calling
spawning procedures in simple_fciqmc, leading to the row of the
Hamiltonian matrix being passed in with size 0 if using -O3.  (Debugging
options prevented this from happening.) Commenting out the call
associated with spawning using a sparse Hamiltonian (which was dead code
for the calculation of interest), leading me to suspect an incorrect
optimisation (perhaps some loop hoisting?).

Inserting an additional test to ensure the components of hamil_csr are
not accessed if they haven't been allocated appears to resolve this.
Note that the additional test does not change the formal logic but makes
the link between hamil_csr being used and sparse_hamil being true clear
to the compiler, preventing such misunderstandings.