More helpful error message if sys argument is not supplied.
We didn't check that a function was called with a table as the argument.
This meant that doing (e.g.)

fci()

resulted in the message

ERROR.
HANDE stops in subroutine: run_lua_hande.
Reason: attempt to index a nil value
EXITING...

which isn't so helpful.

Instead give

ERROR.
HANDE stops in subroutine: get_sys_t.
Reason: No system object supplied.
EXITING...

which hopefully is more helpful.