Modified fix to avoid HDF5 temp array creation
The crashes were being caused by a temporary array being created
when an array slice of the psip arrays was being used.  This
should be optimized away, but in some ifort 13 it wasn't.
The temporary array was placed on the stack, and caused the stack
to be filled entirely (revealed by e.g. ulimit -s 50 before running).
The fix modifies the syntax to not use an array slice, but keeps the
same functionality.