Modify death on composite clusters in CCMC.
Similar to the previous changes to linked death.  If the excip
distribution is correct, then the average change in population on an
excitor at each iteration should be proportional to the population on
that excitor; this is only true, however, when the shift is correct.
Using the projected energy rather than the shift for death on composite
clusters corrects this. The discrepancy is rather smaller than for the
linked algorithm.