Correct parsing of the input file with MPI.
81aec5f2cc accidentally broke this due to the (mistaken) belief that
only the parent processor read the input file.  This was certainly the
intention (given the parent processor read the input file into a buffer
and broadcast it) but then we used open_config_file from AOTUS, which
reads in the input file on all processors.  This failed as the input
filename is now only defined on the parent processor.  Worse, as I had
also changed a lua error to only trigger a stop_all on the parent
processor, a deadlock resulted (as the parent processor was fine...).

The simple fix is to do what was intended: use open_config_chunk and the
(carefully constructed!) buffer containing the input file contents.

NOTE: despite the comments in 81aec5f2cc, this genuinely removes the
requirement for each processor to have access to the input file.