Update documentation for find_weights and excit_dist options
These changes are made after the changes to these options in the
bug_fix/dmqmc_find_weights branch, recently merged into master
(but whose changes are not currently in this branch).

The documentation was also actually wrong before - it
incorrectly stated that the find_weights option should be used
together with the excit_dist option, but actually these were
always specifically *not* supposed to be used together - there
is a stop_all to prevent it.