Handle multiple calculations of the same type in testcode interface.
The testcode interface creates a dict labelled by the calc type.  This
results in only the final calculation of a given type being checked by
testcode.  Instead, append a index giving the position of the
calculation in the output to the calc type so all calculations have
a unique entry in the dict passed to testcode and hence all calculations
are checked by testcode.