reblock_hande.py works with multiple calculations in one file.
Multiple calculations in the same file only works with a single file
(with more than one file they will be correctly reblocked but all
associated with one filename in the output).

The verbosity levels for one file are now the same as for multiple files
(as we want them for multiple calculations).