Faster finding of orbital a in (i,j)->(a,b) for UEG.
Don't check elements of ueg_ternary_conserve that can't contain the
desired orbital.  Given we count the number of set bits in the bit
string containing the possible values of a, we should use this
information to figure out exactly which element in the bit string
contains the selected a.

On a one-off test for IP-DMQMC, this improves the overall calculation by
~15%.  Very worthwhile...