Don't use size function when calling check_allocate
The array size argument to check_allocate is only used to generate an
error message if the allocation fails. But in that case the value of
size(array) will generally not be correct (usually 0 or the size of a
previous allocation of that array, instead of the requested size) so it
is not a helpful way to call the checking function.