Remove use of uninitialised data in select_cluster.
cluster%excitation_level is only meaningful/set for "allowed" clusters,
but it was being tested anyway.  This doesn't cause a bug, as the only
thing done conditional on the value was allowed (false anyway) being
set to false, but not having spurious errors from valgrind makes finding
real errors easier.