Saner handling of per-processor quantities in energy comms buffer.
Split out per-processor and single-value data into separate enums and
provide a procedure for calculating the space required for each
per-processor quantity and the start positions in the buffer.  This
removes the need for magic hard-coding of offsets in multiple places --
instead just get the information from comm_processor_indx.