Read/write vary_shift to a restart file.
If a calculation in the variable shift phase instantaneously has
below the target number of particles when a restart file is created,
the restarted calculation will be run with a fixed shift until the
population is above the target again.  But the shift can often be lower
than the correlation energy, meaning the population decays, or only
slightly above, making the growth very slow.  Either behaviour is not
useful.  The easiest way to avoid it is to save vary_shift in the
restart file (as it is part of the calculation state).