Remove duplicates when restarted from middle
Sometimes restart files are written in the middle of an output file and
the calculation does not end cleanly (no new restart files). The
restarted file will overlap with the previous one for some data points.
There are now removed when merging the two files.
Note that the last duplicate is kept.