Hash the right number of bits to assign deterministic states to processors.
assign_particle_processor takes the number of bits to hash, but the
number of integers was being passed instead.  This only causes a problem
if the bit string is more than one integer (as only the first is hashed)
but shows up in the recently added test, as the max/min deterministic
space size per processor changes on reading it back in.