Set vary shift correctly in restart file written with write_shift
Previously we set vary_shift to false in the restart file written with
restart.write_shift=true (i.e. immediately after the shift is turned on) to
allow the population to continue to grow in a subsequent calculation.
This is now superfluous as qmc.vary_shift allows better control.
Further, we were testing whether the shift should be varied differently
at the start of the calculation and during the calculation, so let's
avoid this confusion...

Note: tests updated which made use of the check at the start of the
calculation for whether to vary the shift - some use older restart files
and some had the target and initial populations identical (which might
not be the greatest idea in the world...).