Convert bit string from spawn_data_t correctly (ie warning-free) in DMQMC initialisation
Note that as we guarantee we do

spawn%sdata(:tot_string_len,idet) = cdet%f

or similar, it is fine to do the reverse as sdata must only hold values
in the range of the i0 kind. (This is not the case when we convert
64-bit strings to 32-bit string when restarting...)

When POP_SIZE=64 and DET_SIZE=32, we must guard the reverse with
int(..., i0) to say we're ok with the potential (as far as the compiler
is concerned) conversion error. This is already done in annihilation
procedures.