Only broadcast 2e integrals from the zero-th processor on each node
Only the zero-th processor on each node has a defined intra-node MPI
communicator and so other processors should not attempt MPI calls with
their intra-node communicator; such calls mut be guarded to check the
communicator is not null. We were doing this on other calls but not this
one -- no idea why nor why only with shmem enabled it became a problem;
it should have caused similar MPI crashes before!