2-Body RDM Support for DMQMC and IP-DMQMC
NB1: replica_tricks is not used here. Maybe it should be included in a way different from FCIQMC to cancel some bias.
NB2: to avoid confusion with the original RDM procedures for the truncated Hilbert space, green_function instead of rdm is used as variable name and the interface is put under the operators table.