Delete T2 sampling in mp1.F90, implement psip_list transfer from mp1 to CCMC.
After tijab is generated exactly as the mp1 wavefunction, the original code went on to sample it again. The short comment said it was for normalisation (no further clarification) and generating higher order terms (it does not seem to do that either). Worse the sampling mysteriously led to an (almost) exact halving of the direct MP2 energy. Therefore that has been removed pending comments from James (original author)
The original code also did not contain an interfact to CCMC/FCIQMC calculations. I have decided to just save the tijab (particle_t object) as a lua object and pass it into CCMC (FCIQMC would be the next step), taking care to suppress CCMC initialisation steps that would overwrite the psip distribution.
Compiles but has yet to be tested.